+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # WaterDrop changelog
 
+## 1.3
+- Drop support for Kafka 0.10 in favor of native support for Kafka 0.11.
+- Support message headers
+- `sasl_over_ssl` support
+- Unlock Ruby Kafka + provide support for 0.7 only
+- #60 - Rename listener to StdoutListener
+- Drop support for Kafka 0.10 in favor of native support for Kafka 0.11.
+- Support ruby-kafka 0.7
+- Support message headers
+- `sasl_over_ssl` support
+- `ssl_client_cert_key_password` support
+
 ## 1.2.5
 - Bump deps to match Karafka
 - drop jruby support